CSC470 Group Project Python/Flask Web Application
This was a group project we completed my Senior year! The web application is a to do list that allows users to create tasks that will be saved on the mongo database. This application included registration, login, creating, editing, and deleting tasks. I mainly worked in the app.py section of this project helping set up the database and getting all of our routes to work correctly. Credit to Connor Starrett, Benjamin Batie, and Kalman Lengyel.